Specify colors in theme, not config file
Colors should be specified in the theme, not the config file. The foreground and background colors were already being set in the theme, so I removed them from the config file. I moved the cursor color to the theme. I removed the mouse pointer color from the config file, because it appears not to have any effect (possibly because it's drawn by GTK now, not Emacs).

I also added theme entries for some frame components (modeline and fringe) that I was simply accepting defaults for previously. I set them to match the defaults, but this makes things more explicit and should make it easier to change them in the future.